- refactor: 포스트 카드 분리
